Q96EB6 had bad a single bad output model 2x
Q96EB6 yielded a single bad output pdb file and 4 good ones, in 2 subsequent runs. The bad pdb files were rejected. All of the models from the second run were added here with the exception of the bad model which was replaced uisng the good model from the first run. On 23May2023, a new I-Tasser run was initiated in the hopes of generating 5 good model pdb file outputs. These are NOT those files...